main_ver 0.0.5 home 화면 결제함 dash board 기능 구현 ERP 인사관리 프로젝트

+++ src/main/java/sjht/erp/login/configure/JwtTokenProvider.java
@@ -34,7 +34,7 @@ public class JwtTokenProvider {
     }
 
     //jwt 토큰 생성
-    public String createToken(Long userId, List<String> roles){
+    public String createToken(int userId, List<String> roles){
         Claims claims = Jwts.claims().setSubject(Long.toString(userId));
         claims.put("roles", roles);
         Date now = new Date();
